+++ src/lib.rs
@@ -319,7 +319,6 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
     }
   }
 
-
   /// Creates a single MOC from a list of cone centers and radius.
   /// # Params
   /// * `lon_deg`: cone centers longitude
@@ -349,19 +348,18 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
         "Cone centers coordinates and radii must be contiguous and in standard order.",
       )),
     }
-      .map_err(PyIOError::new_err)?;
+    .map_err(PyIOError::new_err)?;
     #[cfg(target_arch = "wasm32")]
     {
       // Ignore multi-threading in wasm32
       let cone_it = lon
-        .iter().cloned()
+        .iter()
+        .cloned()
         .zip(lat.iter().cloned())
         .zip(r.iter().cloned());
-      U64MocStore::get_global_store().from_small_cones(
-        depth,
-        delta_depth,
-        cone_it,
-      ).map_err(PyIOError::new_err)
+      U64MocStore::get_global_store()
+        .from_small_cones(depth, delta_depth, cone_it)
+        .map_err(PyIOError::new_err)
     }
     #[cfg(not(target_arch = "wasm32"))]
     {
@@ -370,14 +368,13 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
         .unwrap_or_else(|| num_threads().map(|v| v.get()).unwrap_or(8));
       if n_threads == 1 {
         let cone_it = lon
-          .iter().cloned()
+          .iter()
+          .cloned()
           .zip(lat.iter().cloned())
           .zip(r.iter().cloned());
-        U64MocStore::get_global_store().from_small_cones(
-          depth,
-          delta_depth,
-          cone_it,
-        ).map_err(PyIOError::new_err)
+        U64MocStore::get_global_store()
+          .from_small_cones(depth, delta_depth, cone_it)
+          .map_err(PyIOError::new_err)
       } else {
         let pool = rayon::ThreadPoolBuilder::new()
           .num_threads(n_threads)
@@ -386,12 +383,89 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
         pool
           .install(|| {
             let cone_it = lon
-              .par_iter().cloned()
+              .par_iter()
+              .cloned()
               .zip(lat.par_iter().cloned())
               .zip(r.par_iter().cloned());
-            U64MocStore::get_global_store().from_small_cones_par(
+            U64MocStore::get_global_store().from_small_cones_par(depth, delta_depth, cone_it)
+          })
+          .map_err(PyIOError::new_err)
+      }
+    }
+  }
+
+  /// Creates a single MOC from a list of cone centers and radius.
+  /// # Params
+  /// * `lon_deg`: cone centers longitude
+  /// * `lat_deg`: cone centers latitude
+  /// * `radius_deg`: cones radii
+  /// * `depth`: MOCs depth
+  /// * `delta_depth`: precision parameter
+  /// * `n_threads`: number of threads to use (max number of threads if `n_threads=None`.
+  #[pyfn(m)]
+  fn from_large_cones(
+    lon_deg: PyReadonlyArrayDyn<f64>,
+    lat_deg: PyReadonlyArrayDyn<f64>,
+    radius_deg: PyReadonlyArrayDyn<f64>,
+    depth: u8,
+    delta_depth: u8,
+    n_threads: Option<u16>,
+  ) -> PyResult<usize> {
+    // We zip before multi-threading in case `lon_deg.as_slice()` or ``lon_deg.as_slice`` fail
+    // (due to non-contiguous arrays).
+    let (lon, lat, r) = match (
+      lon_deg.as_slice(),
+      lat_deg.as_slice(),
+      radius_deg.as_slice(),
+    ) {
+      (Ok(lon), Ok(lat), Ok(r)) => Ok((lon, lat, r)),
+      _ => Err(String::from(
+        "Cone centers coordinates and radii must be contiguous and in standard order.",
+      )),
+    }
+    .map_err(PyIOError::new_err)?;
+    #[cfg(target_arch = "wasm32")]
+    {
+      // Ignore multi-threading in wasm32
+      let cone_it = lon
+        .iter()
+        .cloned()
+        .zip(lat.iter().cloned())
+        .zip(r.iter().cloned());
+      U64MocStore::get_global_store()
+        .from_large_cones(depth, delta_depth, CellSelection::All, cone_it)
+        .map_err(PyIOError::new_err)
+    }
+    #[cfg(not(target_arch = "wasm32"))]
+    {
+      let n_threads = n_threads
+        .map(|v| v as usize)
+        .unwrap_or_else(|| num_threads().map(|v| v.get()).unwrap_or(8));
+      if n_threads == 1 {
+        let cone_it = lon
+          .iter()
+          .cloned()
+          .zip(lat.iter().cloned())
+          .zip(r.iter().cloned());
+        U64MocStore::get_global_store()
+          .from_large_cones(depth, delta_depth, CellSelection::All, cone_it)
+          .map_err(PyIOError::new_err)
+      } else {
+        let pool = rayon::ThreadPoolBuilder::new()
+          .num_threads(n_threads)
+          .build()
+          .map_err(|e| PyIOError::new_err(e.to_string()))?;
+        pool
+          .install(|| {
+            let cone_it = lon
+              .par_iter()
+              .cloned()
+              .zip(lat.par_iter().cloned())
+              .zip(r.par_iter().cloned());
+            U64MocStore::get_global_store().from_large_cones_par(
               depth,
               delta_depth,
+              CellSelection::All,
               cone_it,
             )
           })
@@ -399,7 +473,6 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
       }
     }
   }
-
 
   /// Create and store a MOC from the given box.
   ///
@@ -597,6 +670,186 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
     }
   }
 
+  /// Create and store the MOC from the union of the the given box parameters and centers.
+  /// Each boxes covers a limited number of HEALpix cells, else use `from_large_boxes`.
+  ///
+  /// # Input
+  /// * `lon` the longitudes of the centers of the boxes, in degrees
+  /// * `la` the latitudes of the centers of the boxes, in degrees
+  /// * `a` the semi-major axis of the boxes (half the box width), in degrees
+  /// * `b` the semi-minor axis of the boxes (half the box height), in degrees
+  /// * `angle` the position angle of the boxes (i.e. the angle between the north and the semi-major axis, east-of-north), in degrees
+  /// * `depth`: the MOCs depth
+  /// * `n_threads`: number of threads to use (max number of threads if `n_threads=None`.
+  ///
+  /// # Output
+  /// - The MOC index in the storage
+  #[pyfn(m)]
+  pub fn from_small_boxes(
+    lon_deg: PyReadonlyArrayDyn<f64>,
+    lat_deg: PyReadonlyArrayDyn<f64>,
+    a: PyReadonlyArrayDyn<f64>,
+    b: PyReadonlyArrayDyn<f64>,
+    pa: PyReadonlyArrayDyn<f64>,
+    depth: u8,
+    n_threads: Option<u16>,
+  ) -> PyResult<usize> {
+    // We zip before multi-threading in case `lon_deg.as_slice()` or ``lon_deg.as_slice`` fail
+    // (due to non-contiguous arrays).
+    let (lon, lat) = match (lon_deg.as_slice(), lat_deg.as_slice()) {
+      (Ok(lon), Ok(lat)) => Ok((lon, lat)),
+      _ => Err(String::from(
+        "Cone centers coordinates must be contiguous and in standard order.",
+      )),
+    }
+    .map_err(PyIOError::new_err)?;
+    let (a, b, pa) = match (a.as_slice(), b.as_slice(), pa.as_slice()) {
+      (Ok(a), Ok(b), Ok(pa)) => Ok((a, b, pa)),
+      _ => Err(String::from(
+        "Cone centers coordinates must be contiguous and in standard order.",
+      )),
+    }
+    .map_err(PyIOError::new_err)?;
+
+    #[cfg(not(target_arch = "wasm32"))]
+    {
+      let n_threads = n_threads
+        .map(|v| v as usize)
+        .unwrap_or_else(|| num_threads().map(|v| v.get()).unwrap_or(8));
+      if n_threads == 1 {
+        let boxes_params_it = lon.iter().cloned().zip(lat.iter().cloned()).zip(
+          a.iter()
+            .cloned()
+            .zip(b.iter().cloned())
+            .zip(pa.iter().cloned()),
+        );
+        U64MocStore::get_global_store()
+          .from_small_boxes(depth, boxes_params_it)
+          .map_err(PyIOError::new_err)
+      } else {
+        let pool = rayon::ThreadPoolBuilder::new()
+          .num_threads(n_threads)
+          .build()
+          .map_err(|e| PyIOError::new_err(e.to_string()))?;
+        pool
+          .install(|| {
+            let boxes_params_it = lon.par_iter().cloned().zip(lat.par_iter().cloned()).zip(
+              a.par_iter()
+                .cloned()
+                .zip(b.par_iter().cloned())
+                .zip(pa.par_iter().cloned()),
+            );
+            U64MocStore::get_global_store().from_small_boxes_par(depth, boxes_params_it)
+          })
+          .map_err(PyIOError::new_err)
+      }
+    }
+    #[cfg(target_arch = "wasm32")]
+    {
+      let boxes_params_it = lon.iter().cloned().zip(lat.iter().cloned()).zip(
+        a.iter()
+          .cloned()
+          .zip(b.iter().cloned())
+          .zip(pa.iter().cloned()),
+      );
+      U64MocStore::get_global_store()
+        .from_small_boxes(depth, boxes_params_it)
+        .map_err(PyIOError::new_err)
+    }
+  }
+
+  /// Create and store the MOC from the union of the the given box parameters and centers.
+  /// Each boxes covers a large number of HEALpix cells, else use `from_small_boxes`.
+  ///
+  /// # Input
+  /// * `lon` the longitudes of the centers of the boxes, in degrees
+  /// * `la` the latitudes of the centers of the boxes, in degrees
+  /// * `a` the semi-major axis of the boxes (half the box width), in degrees
+  /// * `b` the semi-minor axis of the boxes (half the box height), in degrees
+  /// * `angle` the position angle of the boxes (i.e. the angle between the north and the semi-major axis, east-of-north), in degrees
+  /// * `depth`: the MOCs depth
+  /// * `n_threads`: number of threads to use (max number of threads if `n_threads=None`.
+  ///
+  /// # Output
+  /// - The MOC index in the storage
+  #[pyfn(m)]
+  pub fn from_large_boxes(
+    lon_deg: PyReadonlyArrayDyn<f64>,
+    lat_deg: PyReadonlyArrayDyn<f64>,
+    a: PyReadonlyArrayDyn<f64>,
+    b: PyReadonlyArrayDyn<f64>,
+    pa: PyReadonlyArrayDyn<f64>,
+    depth: u8,
+    n_threads: Option<u16>,
+  ) -> PyResult<usize> {
+    // We zip before multi-threading in case `lon_deg.as_slice()` or ``lon_deg.as_slice`` fail
+    // (due to non-contiguous arrays).
+    let (lon, lat) = match (lon_deg.as_slice(), lat_deg.as_slice()) {
+      (Ok(lon), Ok(lat)) => Ok((lon, lat)),
+      _ => Err(String::from(
+        "Cone centers coordinates must be contiguous and in standard order.",
+      )),
+    }
+    .map_err(PyIOError::new_err)?;
+    let (a, b, pa) = match (a.as_slice(), b.as_slice(), pa.as_slice()) {
+      (Ok(a), Ok(b), Ok(pa)) => Ok((a, b, pa)),
+      _ => Err(String::from(
+        "Cone centers coordinates must be contiguous and in standard order.",
+      )),
+    }
+    .map_err(PyIOError::new_err)?;
+
+    #[cfg(not(target_arch = "wasm32"))]
+    {
+      let n_threads = n_threads
+        .map(|v| v as usize)
+        .unwrap_or_else(|| num_threads().map(|v| v.get()).unwrap_or(8));
+      if n_threads == 1 {
+        let boxes_params_it = lon.iter().cloned().zip(lat.iter().cloned()).zip(
+          a.iter()
+            .cloned()
+            .zip(b.iter().cloned())
+            .zip(pa.iter().cloned()),
+        );
+        U64MocStore::get_global_store()
+          .from_large_boxes(depth, CellSelection::All, boxes_params_it)
+          .map_err(PyIOError::new_err)
+      } else {
+        let pool = rayon::ThreadPoolBuilder::new()
+          .num_threads(n_threads)
+          .build()
+          .map_err(|e| PyIOError::new_err(e.to_string()))?;
+        pool
+          .install(|| {
+            let boxes_params_it = lon.par_iter().cloned().zip(lat.par_iter().cloned()).zip(
+              a.par_iter()
+                .cloned()
+                .zip(b.par_iter().cloned())
+                .zip(pa.par_iter().cloned()),
+            );
+            U64MocStore::get_global_store().from_large_boxes_par(
+              depth,
+              CellSelection::All,
+              boxes_params_it,
+            )
+          })
+          .map_err(PyIOError::new_err)
+      }
+    }
+    #[cfg(target_arch = "wasm32")]
+    {
+      let boxes_params_it = lon.iter().cloned().zip(lat.iter().cloned()).zip(
+        a.iter()
+          .cloned()
+          .zip(b.iter().cloned())
+          .zip(pa.iter().cloned()),
+      );
+      U64MocStore::get_global_store()
+        .from_large_boxes(depth, CellSelection::All, boxes_params_it)
+        .map_err(PyIOError::new_err)
+    }
+  }
+
   /// Create and store a MOC from the given zone.
   ///
   /// # Input
@@ -609,12 +862,24 @@ fn mocpy(m: &Bound<'_, PyModule>) -> PyResult<()> {
   /// # Output
   /// - The index in the storage
   #[pyfn(m)]
-  pub fn from_zone(lon_min: f64, lat_min: f64, lon_max: f64, lat_max: f64, depth: u8) -> PyResult<usize> {
+  pub fn from_zone(
+    lon_min: f64,
+    lat_min: f64,
+    lon_max: f64,
+    lat_max: f64,
+    depth: u8,
+  ) -> PyResult<usize> {
     U64MocStore::get_global_store()
-      .from_zone(lon_min, lat_min, lon_max, lat_max, depth, CellSelection::All)
+      .from_zone(
+        lon_min,
+        lat_min,
+        lon_max,
+        lat_max,
+        depth,
+        CellSelection::All,
+      )
       .map_err(PyValueError::new_err)
   }
-
 
   /// Create a spatial coverage from a given ring.
   ///